kopmall/urls.py
"""
URL configuration for kopmall project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/5.0/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path

from ninja import NinjaAPI
from ninja import Swagger
from ninja_jwt.routers.obtain import obtain_pair_router
from ninja_extra import exceptions

from accounts.api import router as accounts_router

api = NinjaAPI(docs=Swagger())

api.add_router("/users/", accounts_router)
api.add_router("/token", tags=['Auth'], router=obtain_pair_router)

urlpatterns = [
    path('admin/', admin.site.urls),
    path("api/", api.urls)
]


def api_exception_handler(request, exc):
    headers = {}

    if isinstance(exc.detail, (list, dict)):
        data = exc.detail
    else:
        data = {"detail": exc.detail}

    response = api.create_response(request, data, status=exc.status_code)
    for k, v in headers.items():
        response.setdefault(k, v)

    return response

api.exception_handler(exceptions.APIException)(api_exception_handler)