requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.16.0
contextlib2==21.6.0
cryptography==42.0.8
Django==5.0.7
django-lifecycle==1.2.4
django-ninja==1.2.1
django-ninja-extra==0.21.1
django-ninja-jwt==5.3.2
dnspython==2.6.1
email_validator==2.2.0
idna==3.7
injector==0.22.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
ruff==0.5.2
sqlparse==0.5.1
typing_extensions==4.12.2
